refactor: query-builder: Abstract duplicate code from Delete/Update/SoftDelete
* DeleteQueryBuilder, UpdateQueryBuilder and SoftDeleteQueryBuilder abstracted to ModificationQueryBuilder
* SoftDeleteQueryBuilder extends UpdateQueryBuilder
