+++ src/query-builder/builder/DeleteQueryBuilder.ts
@@ -1,49 +1,25 @@
 import {CockroachDriver} from "../../driver/cockroachdb/CockroachDriver";
 import {OracleDriver} from "../../driver/oracle/OracleDriver";
-import {QueryBuilder} from "./QueryBuilder";
-import {ObjectLiteral} from "../../common/ObjectLiteral";
 import {EntityTarget} from "../../common/EntityTarget";
-import {Connection} from "../../connection/Connection";
-import {QueryRunner} from "../../query-runner/QueryRunner";
 import {SqlServerDriver} from "../../driver/sqlserver/SqlServerDriver";
 import {PostgresDriver} from "../../driver/postgres/PostgresDriver";
-import {WhereExpression} from "../WhereExpression";
-import {Brackets} from "../Brackets";
 import {DeleteResult} from "../result/DeleteResult";
-import {ReturningStatementNotSupportedError} from "../../error/ReturningStatementNotSupportedError";
 import {SqljsDriver} from "../../driver/sqljs/SqljsDriver";
 import {MysqlDriver} from "../../driver/mysql/MysqlDriver";
 import {BroadcasterResult} from "../../subscriber/BroadcasterResult";
 import {EntitySchema} from "../../index";
 import {AuroraDataApiDriver} from "../../driver/aurora-data-api/AuroraDataApiDriver";
 import {BetterSqlite3Driver} from "../../driver/better-sqlite3/BetterSqlite3Driver";
+import {ModificationQueryBuilder} from "./ModificationQueryBuilder";
 
 /**
  * Allows to build complex sql queries in a fashion way and execute those queries.
  */
-export class DeleteQueryBuilder<Entity> extends QueryBuilder<Entity> implements WhereExpression {
-
-    // -------------------------------------------------------------------------
-    // Constructor
-    // -------------------------------------------------------------------------
-
-    constructor(connectionOrQueryBuilder: Connection|QueryBuilder<any>, queryRunner?: QueryRunner) {
-        super(connectionOrQueryBuilder as any, queryRunner);
-        this.expressionMap.aliasNamePrefixingEnabled = false;
-    }
+export class DeleteQueryBuilder<Entity> extends ModificationQueryBuilder<Entity> {
 
     // -------------------------------------------------------------------------
     // Public Implemented Methods
     // -------------------------------------------------------------------------
-
-    /**
-     * Gets generated sql query without parameters being replaced.
-     */
-    getQuery(): string {
-        let sql = this.createComment();
-        sql += this.createDeleteExpression();
-        return sql.trim();
-    }
 
     /**
      * Executes sql generated by query builder and returns raw database results.
@@ -141,124 +117,14 @@ export class DeleteQueryBuilder<Entity> extends QueryBuilder<Entity> implements 
         return (this as any) as DeleteQueryBuilder<T>;
     }
 
-    /**
-     * Sets WHERE condition in the query builder.
-     * If you had previously WHERE expression defined,
-     * calling this function will override previously set WHERE conditions.
-     * Additionally you can add parameters used in where expression.
-     */
-    where(where: Brackets|string|((qb: this) => string)|ObjectLiteral|ObjectLiteral[], parameters?: ObjectLiteral): this {
-        this.expressionMap.wheres = []; // don't move this block below since computeWhereParameter can add where expressions
-        const condition = this.computeWhereParameter(where);
-        if (condition)
-            this.expressionMap.wheres = [{ type: "simple", condition: condition }];
-        if (parameters)
-            this.setParameters(parameters);
-        return this;
-    }
-
-    /**
-     * Adds new AND WHERE condition in the query builder.
-     * Additionally you can add parameters used in where expression.
-     */
-    andWhere(where: Brackets|string|((qb: this) => string), parameters?: ObjectLiteral): this {
-        this.expressionMap.wheres.push({ type: "and", condition: this.computeWhereParameter(where) });
-        if (parameters) this.setParameters(parameters);
-        return this;
-    }
-
-    /**
-     * Adds new OR WHERE condition in the query builder.
-     * Additionally you can add parameters used in where expression.
-     */
-    orWhere(where: Brackets|string|((qb: this) => string), parameters?: ObjectLiteral): this {
-        this.expressionMap.wheres.push({ type: "or", condition: this.computeWhereParameter(where) });
-        if (parameters) this.setParameters(parameters);
-        return this;
-    }
-
-    /**
-     * Adds new AND WHERE with conditions for the given ids.
-     */
-    whereInIds(ids: any|any[]): this {
-        return this.where(this.createWhereIdsExpression(ids));
-    }
-
-    /**
-     * Adds new AND WHERE with conditions for the given ids.
-     */
-    andWhereInIds(ids: any|any[]): this {
-        return this.andWhere(this.createWhereIdsExpression(ids));
-    }
-
-    /**
-     * Adds new OR WHERE with conditions for the given ids.
-     */
-    orWhereInIds(ids: any|any[]): this {
-        return this.orWhere(this.createWhereIdsExpression(ids));
-    }
-    /**
-     * Optional returning/output clause.
-     * This will return given column values.
-     */
-    output(columns: string[]): this;
-
-    /**
-     * Optional returning/output clause.
-     * Returning is a SQL string containing returning statement.
-     */
-    output(output: string): this;
-
-    /**
-     * Optional returning/output clause.
-     */
-    output(output: string|string[]): this;
-
-    /**
-     * Optional returning/output clause.
-     */
-    output(output: string|string[]): this {
-        return this.returning(output);
-    }
-
-    /**
-     * Optional returning/output clause.
-     * This will return given column values.
-     */
-    returning(columns: string[]): this;
-
-    /**
-     * Optional returning/output clause.
-     * Returning is a SQL string containing returning statement.
-     */
-    returning(returning: string): this;
-
-    /**
-     * Optional returning/output clause.
-     */
-    returning(returning: string|string[]): this;
-
-    /**
-     * Optional returning/output clause.
-     */
-    returning(returning: string|string[]): this {
-
-        // not all databases support returning/output cause
-        if (!this.connection.driver.isReturningSqlSupported())
-            throw new ReturningStatementNotSupportedError();
-
-        this.expressionMap.returning = returning;
-        return this;
-    }
-
     // -------------------------------------------------------------------------
-    // Protected Methods
+    // Protected Implemented Methods
     // -------------------------------------------------------------------------
 
     /**
      * Creates DELETE express used to perform query.
      */
-    protected createDeleteExpression() {
+    protected createModificationExpression() {
         const tableName = this.getTableName(this.getMainTableName());
         const whereExpression = this.createWhereExpression();
         const returningExpression = this.createReturningExpression();
